Add User, Organisation and Membership tables
* users *

A good rule of thumb is the User model should strictly be used for
authentication and the only associated entities should be specific to
that particular human (e.g. preferences). EVERYTHING else belongs to
Organisation.

$ rails g model User name:string email:string:uniq password_digest:string

- we will use a secure digest to create the password, from the bcrypt
  library
- adds a unique index for the email field so duplicated users cannot be
  created

* organisations *

Member management, roles and permissions and invitations are all essential
features that cannot be based around the User. It's also better to introduce
the concept of an Organisation now rather than have to do it later and
move all the resources associated with a User to the Organisation (which
is a massive job).

$ rails g model Organisation

Business logic for organisations not scoped to this book but the concept is
there for when the times comes (future proofing).

* memberships *

A join table to facilitate the many-to-many relationship between users
and organisations.

$ rails g model Membership user:references organisation:references

Roles can be added to the concept of a Membership but this is out of the
scope of this book.