Extract the page title into a locale dictionary
Defines a helper to format the title

We're going to be using a folder structure for locale dictionaries
based on app/models and app/views. It's common to use this pattern so
that each translation key is scoped to the part of the application
where it is used, e.g. `models.user` or `views.users.show`.

However, something like the page's default title is truly global. Because
of this, we put the translation at the top-level of our locale file rather
than nesting it inside a model- or view-specific namespace.

Using a translation key with a leading dot, automatically scopes it to the
current view.

e.g. t(".title") -> views.controller_name.controller_action.title

Instead we're defining the title at the root level - `en.piazza: Piazza`.

Instead of nesting, similar to:

en:
  activerecord:
    models:
      user: Person
    attributes:
      user:
        name: Name
        email: Email
        password: Password