Add `sign_up` form
Using Bulma CSS (similar to bootstrap) and semantic HTML tags.

HTML supports any custom tag as long as they contain a '-'. The behaviour
is set from the classes used - good practice to prevent a "div soup".

This is currently not working as there are localized strings which do
not belong in a locale dictionary and an undefined `@user` instance
variable.