Add flash message mechanism
`data-turbo-temporary` works to prevent Turbo from caching the alert
during navigation. Without it, users will see the alert for a fraction
of a second when navigating back to a page with a flash after navigating
away due to Turbo's local cache.

* Note on strict locals in paritals *

Added in Rails 7.1, defines local variables used in a partial.

Can be explicitly defined:
   <%#= locals: (name:, position:, last_signed_in:)  -%>

Can use defaults:
   <%#= locals: (name:, position:, last_signed_in: nil)  -%>

Can disallow all local variables:
   <%#= locals: ()  -%>

Great for preventing bugs and as documentation.