Render form input errors
At this point if we enter a value for an attribute of user which fails
validation, and click submit, Rails will wrap the fields for the attributes
in question in the following (e.g. password):

```html
<div class="block">
  <div class="field_with_errors">
    <label class="label" for="user_password">Password</label>
  </div>
  <div class="field_with_errors">
    <input class="input" type="password" name="user[password]" id="user_password">
  </div>
</div>
```

Rendering errors in this way means nothing to Bulma so we need to go in
and modify the behaviour for form errors. `ActionView::Base` has a
`field_error_proc` which can be used to achieve this. Two arguments are
passed to it which are also passed to the form_errors partial, `html_tag`
and `instance`.

`html_tag` is the complete HTML tag for the field which has an error.
`instance` is an instance of the field from the `FormBuilder`, which is a
subclass of the ActionView::Helpers::Tags::Base.

The proc is called for the problematic attributes `<label>` and `<input>`.
We only need to render an error on the input so ignore the label with a
regex.

We use Nokogiri to parse and manipulate the HTML, in this case, adding
the class "is-danger" for Bulma to render the error state.

We also need to render the error message which can be found via the
instance.

* Understanding the `<%==` ERB tag *

Using this tag for `<%== field_html  %>` prevents the HTML being escaped
so it executes and doesn't render on the screen as code.

`<input type="text" name="user[name]" />`

passed to `<%=` will be escaped as:

&lt;input type=&quot;text&quot; name=&quot;user[name]&quot; /&gt;

and the browser will display it as the code above.

`<input type="text" name="user[name]" />` passed to `<%==` will output:

`<input type="text" name="user[name]" />` and the browser will render
the input field rather than showing the tags as text.

* Going into the Rails source code *

Each time a form helper method (like f.text_field :name) generates an HTML tag,
it goes through an internal “tag builder” class. When that tag builder detects
the presence of an error on the corresponding object attribute, it hands the
generated HTML string (the “tag”) plus metadata (the “instance”) over to
field_error_proc.

actiontext/app/helpers/action_text/tag_helper.rb

def error_wrapping(html_tag)
  if object_has_errors?
    @template_object.instance_exec(html_tag, self, &Base.field_error_proc)
  else
    html_tag
  end
end

- `@template_object` is typically the view context
- `instance_exec(html_tag, self, &Base.field_error_proc)` means:
   - call the field_error_proc (`&Base.field_error_proc`)
   - within the execution context of `@template_object`
   - passing `html_tag` and `self` as arguments

So... `Proc.new { |html_tag, instance| ... }`

- First argument: html_tag (the raw HTML for your form field, e.g. "<input ...>")
- Second argument: instance (the tag builder object itself — in this case,
  it's an instance of ActionView::Helpers::Tags::Base or one of its subclasses,
  such as ActionText)

Passing self into the field_error_proc is effectively passing “the tag builder instance”
as the second argument.

Proc.new do |html_tag, instance|
  # instance.object       => the model object (e.g., @post, @user, etc.)
  # instance.method_name  => the attribute name (e.g., :title, :name, etc.)
  # instance.error_message => the validation errors for this field
  # ...
end