Add sessions_controller#create
Creates a session record, calling the class method we just defined on User

From a user's point of view, they're logged in. This is not true of the
App's. Added a 'TODO:' and will cover this next.

** Notes **

Notice the use of `flash.now[:danger]` (red alert message)

I can see that the error messages stack on top of eachother which will
need to be fixed in this PR?

Interesting use of conditional assignment for the `permitted_params`.

DEPRECATION WARNING: Passing nil to the :model argument is deprecated and will raise in Rails 8.0 (called from _app_views_sessions_new_html_erb___1135633806037528300_21300 at /Users/oliveratkins/code/piazza-web/app/views/sessions/new.html.erb:9)