Add Current class for global attributes on every request
** Notes on Current **

Defines thread-isolated attribute which are accessible across models,
controllers and services and reset at the end of each request.

Need to exercise extreme caution:

“A word of caution: It's easy to overdo a global singleton like Current
and tangle your model as a result. Current should only be used for a few,
top-level globals, like account, user, and request details. The attributes
stuck in Current should be used by more or less all actions on all requests.”

A user undeniably qualifies a global.