Add authentication methods to Authentication concern
** #authenticate **

Assigns `Current.session` and `Current.user` if the user is successfully
authenticated. If more methods of authentication need to be added, they
can with a logical OR operator:

  def authenticate
    Current.session = authenticate_using_cookie ||
      authenticate_using_token

    Current.user = Current.session&.user
  end

** #authenticate_using_cookie **

Extracts the data from the encrypted cookie and passes to the
`authenticate_using method` method. It converts the session to a hash
with indifferent access. This means values can be retrieved using both
the symbol and string versions of the key.

** #authenticate_using **

First line deconstructs the hash into variables using the Ruby 3's rightward
assignment syntax (fancy!).

If the hash cannot be deconstructed using the symbols specified, a
NoMatchingPatternError is thrown, essentially meaning authentication was
unsuccessful. Error rescued and nil returned.

Finally, user object used to authenticate the session

user.authenticate_session(session, token)

which currently will not work as there is no #authenticate_session
method on user (to be added next!).